+++ front/react-front-project/src/api/productApi.js
@@ -5,3 +5,13 @@ export async function getProducts(categoryId) {
   const response = await axios.get('/api/products', { params })
   return response.data
 }
+
+export async function getProduct(productId) {
+  const response = await axios.get(`/api/products/${productId}`)
+  return response.data
+}
+
+export async function createProduct(productData){
+  const response = await axios.post('/api/products', productData)
+  return response.data
+}